Define mockable traits in mqtt library
