+++ common/mqtt_client/examples/sawtooth_publisher.rs
@@ -4,10 +4,7 @@ use futures_timer::Delay;
 use log::debug;
 use log::error;
 use log::info;
-use mqtt_client::Config;
-use mqtt_client::Message;
-use mqtt_client::Topic;
-use mqtt_client::{Client, ErrorStream, MessageStream};
+use mqtt_client::{Client, Config, Message, MqttClient, MqttErrorStream, MqttMessageStream, Topic};
 use std::convert::TryFrom;
 use std::env;
 use std::io::Write;
@@ -179,7 +176,7 @@ async fn publish_multi_topic(
     Ok(())
 }
 
-async fn listen_command(mut messages: MessageStream) {
+async fn listen_command(mut messages: Box<dyn MqttMessageStream>) {
     while let Some(message) = messages.next().await {
         debug!("C8Y command: {:?}", message.payload);
         if let Some(cmd) = std::str::from_utf8(&message.payload).ok() {
@@ -191,7 +188,7 @@ async fn listen_command(mut messages: MessageStream) {
     }
 }
 
-async fn listen_c8y_error(mut messages: MessageStream) {
+async fn listen_c8y_error(mut messages: Box<dyn MqttMessageStream>) {
     let mut count: u32 = 0;
     while let Some(message) = messages.next().await {
         error!("C8Y error: {:?}", message.payload);
@@ -202,7 +199,7 @@ async fn listen_c8y_error(mut messages: MessageStream) {
     }
 }
 
-async fn listen_error(mut errors: ErrorStream) {
+async fn listen_error(mut errors: Box<dyn MqttErrorStream>) {
     let mut count: u32 = 0;
     while let Some(error) = errors.next().await {
         error!("System error: {}", error);
