+++ common/mqtt_client/examples/temperature_publisher.rs
@@ -4,10 +4,7 @@ use futures_timer::Delay;
 use log::debug;
 use log::error;
 use log::info;
-use mqtt_client::Config;
-use mqtt_client::Message;
-use mqtt_client::Topic;
-use mqtt_client::{Client, ErrorStream, MessageStream};
+use mqtt_client::{Client, Config, Message, MqttClient, MqttErrorStream, MqttMessageStream, Topic};
 use rand::prelude::*;
 use std::time::Duration;
 
@@ -63,7 +60,7 @@ fn random_in_range(low: i32, high: i32) -> i32 {
     rng.gen_range(low..high)
 }
 
-async fn listen_command(mut messages: MessageStream) {
+async fn listen_command(mut messages: Box<dyn MqttMessageStream>) {
     while let Some(message) = messages.next().await {
         debug!("C8Y command: {:?}", message.payload);
         if let Some(cmd) = std::str::from_utf8(&message.payload).ok() {
@@ -75,13 +72,13 @@ async fn listen_command(mut messages: MessageStream) {
     }
 }
 
-async fn listen_c8y_error(mut messages: MessageStream) {
+async fn listen_c8y_error(mut messages: Box<dyn MqttMessageStream>) {
     while let Some(message) = messages.next().await {
         error!("C8Y error: {:?}", message.payload);
     }
 }
 
-async fn listen_error(mut errors: ErrorStream) {
+async fn listen_error(mut errors: Box<dyn MqttErrorStream>) {
     while let Some(error) = errors.next().await {
         error!("System error: {}", error);
     }
